refactor(api): api 통신 관련 설정 파일 경로 변경
