+++ src/api/components/ThemeErrorFallback.tsx
@@ -16,8 +16,10 @@ export const ThemeErrorFallback = ({ error }: ErrorFallbackProps) => {
   useEffect(() => {
     if (error.message === RENDER_ERROR_MESSAGES.THEME_NOT_FOUND) {
       navigate(ROUTER_PATH.HOME);
+    } else {
+      throw error;
     }
-  }, [error.message, navigate]);
+  }, [error, navigate]);
 
-  return <div>error page</div>;
+  return <div>{error.message}</div>;
 };
