+++ src/api/config/errorHandler.ts
@@ -1,0 +1,20 @@
+import { AxiosError } from 'axios';
+
+import { API_ERROR_MESSAGES } from '@/constants/errorMessage';
+
+export function getErrorMessage(error: AxiosError): string {
+  if (error.response) {
+    switch (error.response.status) {
+      case 404:
+        return API_ERROR_MESSAGES.DATA_NOT_FOUND;
+      default:
+        return API_ERROR_MESSAGES.FETCH_ERROR;
+    }
+  }
+
+  if (error.request) {
+    return API_ERROR_MESSAGES.NETWORK_ERROR;
+  }
+
+  return API_ERROR_MESSAGES.UNKNOWN_ERROR;
+}
