+++ src/api/config/instance.ts
@@ -1,0 +1,15 @@
+import axios, { AxiosInstance, AxiosRequestConfig } from 'axios';
+
+export const initInstance = (config: AxiosRequestConfig): AxiosInstance => {
+  const instance = axios.create({
+    timeout: 5000,
+    ...config,
+    headers: {
+      Accept: 'application/json',
+      'Content-Type': 'application/json',
+      ...config.headers,
+    },
+  });
+
+  return instance;
+};
